Remove dependency on matsim run output config file
The keys in the matsim config file change over the time and become deprecated or defunct. Since this requires to have a reader from the specific matsim version, incompatibilities arise. To avoid this the dependency is removed, the only downside is that now the network CRS has to be provided as further input.